requirements.txt
fastapi
uvicorn
pydantic
python-decouple
pymongo